feat(backend): store money as integer minor units for exact sums
Amounts are persisted as INTEGER minor units (cents/satoshis) and converted to
major units only at the API boundary, eliminating float drift (ten $0.10 items
now sum to exactly 1.00). Aggregations group by currency, so per-currency totals
are exact and USD/PLN/BTC are never summed into one number.

+++ backend/src/config/database.ts
@@ -31,7 +31,7 @@ export function initializeDatabase(): void {
   const createTableSQL = `
     CREATE TABLE IF NOT EXISTS expenses (
       id INTEGER PRIMARY KEY AUTOINCREMENT,
-      amount REAL NOT NULL CHECK(amount > 0),
+      amount INTEGER NOT NULL CHECK(amount > 0),
       date TEXT NOT NULL,
       description TEXT NOT NULL CHECK(length(description) > 0),
       category TEXT NOT NULL CHECK(category IN ('groceries', 'transport', 'media', 'entertainment', 'utilities', 'maintenance', 'other')),
@@ -82,7 +82,7 @@ export function initializeDatabase(): void {
         db.exec(`
           CREATE TABLE expenses_new (
             id INTEGER PRIMARY KEY AUTOINCREMENT,
-            amount REAL NOT NULL CHECK(amount > 0),
+            amount INTEGER NOT NULL CHECK(amount > 0),
             date TEXT NOT NULL,
             description TEXT NOT NULL CHECK(length(description) > 0),
             category TEXT NOT NULL CHECK(category IN ('groceries', 'transport', 'media', 'entertainment', 'utilities', 'maintenance', 'other')),
@@ -137,7 +137,7 @@ export function initializeDatabase(): void {
         db.exec(`
           CREATE TABLE expenses_new (
             id INTEGER PRIMARY KEY AUTOINCREMENT,
-            amount REAL NOT NULL CHECK(amount > 0),
+            amount INTEGER NOT NULL CHECK(amount > 0),
             date TEXT NOT NULL,
             description TEXT NOT NULL CHECK(length(description) > 0),
             category TEXT NOT NULL CHECK(category IN ('groceries', 'transport', 'media', 'entertainment', 'utilities', 'maintenance', 'other')),
